Support Java 7 without sending NaN.
A recent commit added a check for infinite numbers and NaN with a Java 8
API (Double.isfinite()). This was changed in a subsequent commit to a Java 7
API (Double.isInfinite()) but this does not check for NaN.

To support Java 7 and not send NaN an extra check using Double.isNaN() is
required.